use dependancy injection for service graph construction
The executor could not easily be extended with functionality not defined
in the go-service package. This commit changes this. Instead of using
direct dependancies, dependancy injection is introduced. Users of
go-service can create services executed by the executor which depend on
plain struct or services from other packages by using the service 
registry.